+++ src/constants.ts
@@ -74,6 +74,7 @@ export const TEST_IDS = {
     fieldFilterMode: createSelector('data-testid column-editor field-filter-mode'),
     fieldFilterVariable: createSelector('data-testid column-editor field-filter-variable'),
     fieldSortEnabled: createSelector('data-testid column-editor field-sort-enabled'),
+    fieldPinEnabled: createSelector('data-testid column-editor field-pin-enabled'),
     fieldAppearanceBackgroundApplyToRow: createSelector(
       'data-testid column-editor field-appearance-background-apply-to-row'
     ),